Move help functions to another file and better error handling
